[feat] Add full compaction logic involing > 2 levels
This commit adds all the required logic to introduce a compaction that involves
all levels L0-Ln. The allocation log (manifest) does not work yet and the system
is not recoverable. However, we do not lose data when the new compaction
mechanism is triggered and YCSB A-D workloads pass. When the new compaction is
triggered we make the last level size to be infinite to keep ingestion going.
When split is introduced this will be removed